Adapted data types for 64 bit and improved convertTo functions

+++ backend/include/configfile.hpp
@@ -88,7 +88,7 @@ public:
 		/**
 		  * The tDisk's blocksize
 		 **/
-		long blocksize;
+		unsigned int blocksize;
 
 		/**
 		  * The internal devices of the tDisk
